always remove message handler on platform thread
we do not know which thread deinit will be called on, so ensure it is called on
the (main) platform thread by spawning a task. Some refactoring was required
because deinit cannot capture self in Swift 6 language mode.

For clarity, this is open coded in each deinit to avoid suppressing this
warning.

Fixes: #16